moved the last four functions
